+++ Source/XsollaUtils/Private/XsollaDemoGameModeBase.cpp
@@ -1,0 +1,13 @@
+// Copyright 2020 Xsolla Inc. All Rights Reserved.
+
+#include "XsollaDemoGameModeBase.h"
+
+TSubclassOf<UUserWidget> AXsollaDemoGameModeBase::GetLoginDemo() const
+{
+	return LoginDemo;
+}
+
+TSubclassOf<UUserWidget> AXsollaDemoGameModeBase::GetStoreDemo() const
+{
+	return StoreDemo;
+}