+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 2020-12-25
+
+### Changed
+- Authentication via access token with custom parameters.
+- New web store for demo project.
+
 ## [0.1.0] - 2020-11-12
 
 ### Added